chore(release): 2.11.1
Initial pass at reorganization

Reorginanization of the repository so that it is code focused vs. being
focused on how the packages are built. Packaging will be moved to
GitHub Actions so that every commit will create new packages that can
be used for testing. Releases will be created and produce packages
when a tag is committed to the repository.

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

Added beginnings of package building code

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

Updated GitHub workflow

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

Storing generated package as an artifact

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

Refactored build API workflow

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

More reorganization of files

Moved MeshChat source to top level

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

feat: Set send channel when channel filter changed

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

docs: Updated README with history of MeshChat

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

fix: set epoch in send_message API even if not specified

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

feat: add message class abstracting message handling (#23)

* refactor: add comments to clarify flow

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

* feat: add messages class to front end

Update the handling of messages on the front end code to better control
the update of messages and coordinateion with the rest of the UI.

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

* add more config support

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

* feat: message ID can be specified in send_message API

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

* add md5 module to front-end

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

* integrate message class into front-end code

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

---------

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

Add GitHub Actions workflow to automate releases (#40)

* Added beginnings of package building code
* chore: create release CI job
* chore: add documentation section to changelog
* chore: make build-meshchat-package a reusable workflow
* chore: add build packages to release
* fix: update-version.sh can now be told version to set
* chore(ci): convert meshchat-api build to reusable workflow
* chore(ci): set release to only work on release branch

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

chore(ci): update release pipeline

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>

feat: allow admin to set default channel (#19)

A MeshChat administrator now has the ability to define the default
channel that a newly logged in user is placed into. This requires that
the administrator set the value `default_channel` in the configuration
file. Setting `default_channel` to an empty string ('') will continue
using the legacy setting of 'Everything'.

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>