Add support for meshchat_local.lua
Package upgrades do not update the `meshchatconfig.lua` because
it is tagged as a configuration file. This prevents settings
from being lost, but makes it difficult to add new settings
in future releases. This change establishes the `meshchat_local.lua`
file for local administrators to configure there MeshChat.

During installation of a package if a `meshchat_local.lua` is not
found, then a copy of `meshchatconfig.lua` is copied and
sanitized of the `app_version` and `protocol_version` settings.

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>